Defined unsigned long long as UInt64 instead

+++ AltNum/AltDec/AltDec.hpp
@@ -750,14 +750,14 @@ public:
         constexpr auto PartialUIntDivOpV1 = PartialUIntDivOpV1<unsigned int>;
         constexpr auto PartialIntDivOpV1 = PartialIntDivOpV1<signed int>;
         constexpr auto UnsignedPartialIntDivOpV1 = PartialUIntDivOpV1<signed int>;
-        constexpr auto PartialUInt64DivOpV1 = PartialUIntDivOpV1<unsigned long long>;
-        constexpr auto PartialInt64DivOpV1 = PartialIntDivOpV1<signed long long>;
+        constexpr auto PartialUInt64DivOpV1 = PartialUIntDivOpV1<UInt64>;
+        constexpr auto PartialInt64DivOpV1 = PartialIntDivOpV1<Int64>;
 
         constexpr auto BasicUIntDivOp = BasicUIntDivOpV1<unsigned int>;
         constexpr auto BasicIntDivOp = BasicIntDivOpV1<signed int>;
         constexpr auto UnsignedBasicIntDivOp = BasicUIntDivOpV1<signed int>;
-        constexpr auto BasicUInt64DivOp = BasicUIntDivOpV1<unsigned long long>;
-        constexpr auto BasicInt64DivOp = BasicIntDivOpV1<signed long long>;
+        constexpr auto BasicUInt64DivOp = BasicUIntDivOpV1<UInt64>;
+        constexpr auto BasicInt64DivOp = BasicIntDivOpV1<Int64>;
         constexpr auto UnsignedBasicInt64DivOp = BasicUIntDivOpV1<signed int>;
 
         constexpr auto BasicUInt8DivOp = BasicUIntDivOpV1<unsigned char>;
@@ -768,8 +768,8 @@ public:
         constexpr auto BasicDivideByUInt = BasicDivideByUIntV1<unsigned int>;
         constexpr auto BasicDivideByInt = BasicDivideByIntV1<signed int>;
         constexpr auto UnsignedBasicDivideByInt = BasicDivideByUIntV1<signed int>;
-        constexpr auto BasicDivideByUInt64 = BasicDivideByUIntV1<unsigned long long>;
-        constexpr auto BasicDivideByInt64 = BasicDivideByIntV1<signed long long>;
+        constexpr auto BasicDivideByUInt64 = BasicDivideByUIntV1<UInt64>;
+        constexpr auto BasicDivideByInt64 = BasicDivideByIntV1<Int64>;
         constexpr auto UnsignedBasicDivideByInt64 = BasicDivideByUIntV1<signed int>;
 
         constexpr auto BasicDivideByUInt8 = BasicDivideByUIntV1<unsigned char>;
@@ -947,14 +947,14 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec operator/(const AltDec& self, const signed int& Value) { return self.DivideByInt(Value); }
-        friend AltDec operator/(const AltDec& self, const signed long long& Value) { return self.DivideByInt64(Value); }
+        friend AltDec operator/(const AltDec& self, const Int64& Value) { return self.DivideByInt64(Value); }
         friend AltDec operator/(const AltDec& self, const unsigned int& Value) { return self.DivideByUInt(Value); }
-        friend AltDec operator/(const AltDec& self, const unsigned long long& Value) { return self.DivideByUInt64(Value); }
+        friend AltDec operator/(const AltDec& self, const UInt64& Value) { return self.DivideByUInt64(Value); }
 		
         friend AltDec operator/(const signed int& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
-        friend AltDec operator/(const signed long long& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
+        friend AltDec operator/(const Int64& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
         friend AltDec operator/(const unsigned int& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
-        friend AltDec operator/(const unsigned long long& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
+        friend AltDec operator/(const UInt64& lValue, const AltDec& rValue) { return ((AltDec)lValue).DivideBy(rValue); }
 
         friend AltDec operator/(const AltDec& self, const signed char& Value) { return self.DivideByInt8(Value); }
         friend AltDec operator/(const AltDec& self, const signed short& Value) { return self.DivideByInt16(Value); }
@@ -974,9 +974,9 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec& operator/=(AltDec& self, const signed int& Value) { return self.IntDivOp(Value); }
-        friend AltDec& operator/=(AltDec& self, const signed long long& Value) { return self.Int64DivOp(Value); }
+        friend AltDec& operator/=(AltDec& self, const Int64& Value) { return self.Int64DivOp(Value); }
         friend AltDec& operator/=(AltDec& self, const unsigned int& Value) { return self.UIntDivOp(Value); }
-        friend AltDec& operator/=(AltDec& self, const unsigned long long& Value) { return self.UInt64DivOp(Value); }
+        friend AltDec& operator/=(AltDec& self, const UInt64& Value) { return self.UInt64DivOp(Value); }
 
         friend AltDec& operator/=(AltDec& self, const signed char& Value) { return self.Int8DivOp(Value); }
         friend AltDec& operator/=(AltDec& self, const signed short& Value) { return self.Int16DivOp(Value); }
@@ -1037,14 +1037,14 @@ public:
         constexpr auto PartialUIntMultOpV1 = PartialUIntMultOpV1<unsigned int>;
         constexpr auto PartialIntMultOpV1 = PartialIntMultOpV1<signed int>;
         constexpr auto UnsignedPartialIntMultOpV1 = PartialUIntMultOpV1<signed int>;
-        constexpr auto PartialUInt64MultOpV1 = PartialUIntMultOpV1<unsigned long long>;
-        constexpr auto PartialInt64MultOpV1 = PartialIntMultOpV1<signed long long>;
+        constexpr auto PartialUInt64MultOpV1 = PartialUIntMultOpV1<UInt64>;
+        constexpr auto PartialInt64MultOpV1 = PartialIntMultOpV1<Int64>;
 
         constexpr auto BasicUIntMultOp = BasicUIntMultOpV1<unsigned int>;
         constexpr auto BasicIntMultOp = BasicIntMultOpV1<signed int>;
         constexpr auto UnsignedBasicIntMultOp = BasicUIntMultOpV1<signed int>;
-        constexpr auto BasicUInt64MultOp = BasicUIntMultOpV1<unsigned long long>;
-        constexpr auto BasicInt64MultOp = BasicIntMultOpV1<signed long long>;
+        constexpr auto BasicUInt64MultOp = BasicUIntMultOpV1<UInt64>;
+        constexpr auto BasicInt64MultOp = BasicIntMultOpV1<Int64>;
         constexpr auto UnsignedBasicInt64MultOp = BasicUIntMultOpV1<signed int>;
 
         constexpr auto BasicUInt8MultOp = BasicUIntMultOpV1<unsigned char>;
@@ -1055,8 +1055,8 @@ public:
         constexpr auto BasicMultipleByUInt = BasicMultipleByUIntV1<unsigned int>;
         constexpr auto BasicMultipleByInt = BasicMultipleByIntV1<signed int>;
         constexpr auto UnsignedBasicMultipleByInt = BasicMultipleByUIntV1<signed int>;
-        constexpr auto BasicMultipleByUInt64 = BasicMultipleByUIntV1<unsigned long long>;
-        constexpr auto BasicMultipleByInt64 = BasicMultipleByIntV1<signed long long>;
+        constexpr auto BasicMultipleByUInt64 = BasicMultipleByUIntV1<UInt64>;
+        constexpr auto BasicMultipleByInt64 = BasicMultipleByIntV1<Int64>;
         constexpr auto UnsignedBasicMultipleByInt64 = BasicMultipleByUIntV1<signed int>;
 
         constexpr auto BasicMultipleByUInt8 = BasicMultipleByUIntV1<unsigned char>;
@@ -1235,14 +1235,14 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec operator*(const AltDec& self, const signed int& Value) { return self.MultiplyByInt(Value); }
-        friend AltDec operator*(const AltDec& self, const signed long long& Value) { return self.MultiplyByInt64(Value); }
+        friend AltDec operator*(const AltDec& self, const Int64& Value) { return self.MultiplyByInt64(Value); }
         friend AltDec operator*(const AltDec& self, const unsigned int& Value) { return self.MultiplyByUInt(Value); }
-        friend AltDec operator*(const AltDec& self, const unsigned long long& Value) { return self.MultiplyByUInt64(Value); }
+        friend AltDec operator*(const AltDec& self, const UInt64& Value) { return self.MultiplyByUInt64(Value); }
 		
         friend AltDec operator*(const signed int& lValue, const AltDec& rValue) { return rValue.MultiplyByInt(lValue); }
-        friend AltDec operator*(const signed long long& lValue, const AltDec& rValue) { return rValue.MultiplyByInt64(lValue); }
+        friend AltDec operator*(const Int64& lValue, const AltDec& rValue) { return rValue.MultiplyByInt64(lValue); }
         friend AltDec operator*(const unsigned int& lValue, const AltDec& rValue) { return rValue.MultiplyByUInt(lValue); }
-        friend AltDec operator*(const unsigned long long& lValue, const AltDec& rValue) { return rValue.MultiplyByUInt64(lValue); }
+        friend AltDec operator*(const UInt64& lValue, const AltDec& rValue) { return rValue.MultiplyByUInt64(lValue); }
 
         friend AltDec operator*(const AltDec& self, const signed char& Value) { return self.MultiplyByInt8(Value); }
         friend AltDec operator*(const AltDec& self, const signed short& Value) { return self.MultiplyByInt16(Value); }
@@ -1261,9 +1261,9 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec& operator*=(AltDec& self, const signed int& Value) { return self.IntMultOp(Value); }
-        friend AltDec& operator*=(AltDec& self, const signed long long& Value) { return self.Int64MultOp(Value); }
+        friend AltDec& operator*=(AltDec& self, const Int64& Value) { return self.Int64MultOp(Value); }
         friend AltDec& operator*=(AltDec& self, const unsigned int& Value) { return self.UIntMultOp(Value); }
-        friend AltDec& operator*=(AltDec& self, const unsigned long long& Value) { return self.UInt64MultOp(Value); }
+        friend AltDec& operator*=(AltDec& self, const UInt64& Value) { return self.UInt64MultOp(Value); }
 
         friend AltDec& operator*=(AltDec& self, const signed char& Value) { return self.Int8MultOp(Value); }
         friend AltDec& operator*=(AltDec& self, const signed short& Value) { return self.Int16MultOp(Value); }
@@ -1324,14 +1324,14 @@ public:
         constexpr auto PartialUIntAddOpV1 = PartialUIntAddOpV1<unsigned int>;
         constexpr auto PartialIntAddOpV1 = PartialIntAddOpV1<signed int>;
         constexpr auto UnsignedPartialIntAddOpV1 = PartialUIntAddOpV1<signed int>;
-        constexpr auto PartialUInt64AddOpV1 = PartialUIntAddOpV1<unsigned long long>;
-        constexpr auto PartialInt64AddOpV1 = PartialIntAddOpV1<signed long long>;
+        constexpr auto PartialUInt64AddOpV1 = PartialUIntAddOpV1<UInt64>;
+        constexpr auto PartialInt64AddOpV1 = PartialIntAddOpV1<Int64>;
 
         constexpr auto BasicAddByUInt = BasicAddByUIntV1<unsigned int>;
         constexpr auto BasicAddByInt = BasicAddByIntV1<signed int>;
         constexpr auto UnsignedBasicAddByInt = BasicAddByUIntV1<signed int>;
-        constexpr auto BasicAddByUInt64 = BasicAddByUIntV1<unsigned long long>;
-        constexpr auto BasicAddByInt64 = BasicAddByIntV1<signed long long>;
+        constexpr auto BasicAddByUInt64 = BasicAddByUIntV1<UInt64>;
+        constexpr auto BasicAddByInt64 = BasicAddByIntV1<Int64>;
         constexpr auto UnsignedBasicAddByInt64 = BasicAddByUIntV1<signed int>;
 
         constexpr auto BasicAddByUInt8 = BasicAddByUIntV1<unsigned char>;
@@ -1492,14 +1492,14 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec operator+(const AltDec& self, const signed int& Value) { return self.AddByInt(Value); }
-        friend AltDec operator+(const AltDec& self, const signed long long& Value) { return self.AddByInt64(Value); }
+        friend AltDec operator+(const AltDec& self, const Int64& Value) { return self.AddByInt64(Value); }
         friend AltDec operator+(const AltDec& self, const unsigned int& Value) { return self.AddByUInt(Value); }
-        friend AltDec operator+(const AltDec& self, const unsigned long long& Value) { return self.AddByUInt64(Value); }
+        friend AltDec operator+(const AltDec& self, const UInt64& Value) { return self.AddByUInt64(Value); }
 		
         friend AltDec operator+(const signed int& lValue, const AltDec& rValue) { return rValue.AddByInt(lValue); }
-        friend AltDec operator+(const signed long long& lValue, const AltDec& rValue) { return rValue.AddByInt64(lValue); }
+        friend AltDec operator+(const Int64& lValue, const AltDec& rValue) { return rValue.AddByInt64(lValue); }
         friend AltDec operator+(const unsigned int& lValue, const AltDec& rValue) { return rValue.AddByUInt(lValue); }
-        friend AltDec operator+(const unsigned long long& lValue, const AltDec& rValue) { return rValue.AddByUInt64(lValue); }
+        friend AltDec operator+(const UInt64& lValue, const AltDec& rValue) { return rValue.AddByUInt64(lValue); }
 
         friend AltDec operator+(const AltDec& self, const signed char& Value) { return self.AddByInt8(Value); }
         friend AltDec operator+(const AltDec& self, const signed short& Value) { return self.AddByInt16(Value); }
@@ -1518,9 +1518,9 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec& operator+=(AltDec& self, const signed int& Value) { return self.IntAddOp(Value); }
-        friend AltDec& operator+=(AltDec& self, const signed long long& Value) { return self.Int64AddOp(Value); }
+        friend AltDec& operator+=(AltDec& self, const Int64& Value) { return self.Int64AddOp(Value); }
         friend AltDec& operator+=(AltDec& self, const unsigned int& Value) { return self.UIntAddOp(Value); }
-        friend AltDec& operator+=(AltDec& self, const unsigned long long& Value) { return self.UInt64AddOp(Value); }
+        friend AltDec& operator+=(AltDec& self, const UInt64& Value) { return self.UInt64AddOp(Value); }
 
         friend AltDec& operator+=(AltDec& self, const signed char& Value) { return self.Int8AddOp(Value); }
         friend AltDec& operator+=(AltDec& self, const signed short& Value) { return self.Int16AddOp(Value); }
@@ -1581,14 +1581,14 @@ public:
         constexpr auto PartialUIntSubOpV1 = PartialUIntSubOpV1<unsigned int>;
         constexpr auto PartialIntSubOpV1 = PartialIntSubOpV1<signed int>;
         constexpr auto UnsignedPartialIntSubOpV1 = PartialUIntSubOpV1<signed int>;
-        constexpr auto PartialUInt64SubOpV1 = PartialUIntSubOpV1<unsigned long long>;
-        constexpr auto PartialInt64SubOpV1 = PartialIntSubOpV1<signed long long>;
+        constexpr auto PartialUInt64SubOpV1 = PartialUIntSubOpV1<UInt64>;
+        constexpr auto PartialInt64SubOpV1 = PartialIntSubOpV1<Int64>;
 
         constexpr auto BasicUIntSubOp = BasicUIntSubOpV1<unsigned int>;
         constexpr auto BasicIntSubOp = BasicIntSubOpV1<signed int>;
         constexpr auto UnsignedBasicIntSubOp = BasicUIntSubOpV1<signed int>;
-        constexpr auto BasicUInt64SubOp = BasicUIntSubOpV1<unsigned long long>;
-        constexpr auto BasicInt64SubOp = BasicIntSubOpV1<signed long long>;
+        constexpr auto BasicUInt64SubOp = BasicUIntSubOpV1<UInt64>;
+        constexpr auto BasicInt64SubOp = BasicIntSubOpV1<Int64>;
         constexpr auto UnsignedBasicInt64SubOp = BasicUIntSubOpV1<signed int>;
 
         constexpr auto BasicUInt8SubOp = BasicUIntSubOpV1<unsigned char>;
@@ -1599,8 +1599,8 @@ public:
         constexpr auto BasicSubtractByUInt = BasicSubtractByUIntV1<unsigned int>;
         constexpr auto BasicSubtractByInt = BasicSubtractByIntV1<signed int>;
         constexpr auto UnsignedBasicSubtractByInt = BasicSubtractByUIntV1<signed int>;
-        constexpr auto BasicSubtractByUInt64 = BasicSubtractByUIntV1<unsigned long long>;
-        constexpr auto BasicSubtractByInt64 = BasicSubtractByIntV1<signed long long>;
+        constexpr auto BasicSubtractByUInt64 = BasicSubtractByUIntV1<UInt64>;
+        constexpr auto BasicSubtractByInt64 = BasicSubtractByIntV1<Int64>;
         constexpr auto UnsignedBasicSubtractByInt64 = BasicSubtractByUIntV1<signed int>;
 
         constexpr auto BasicSubtractByUInt8 = BasicSubtractByUIntV1<unsigned char>;
@@ -1762,14 +1762,14 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec operator-(const AltDec& self, const signed int& Value) { return self.SubtractByInt(Value); }
-        friend AltDec operator-(const AltDec& self, const signed long long& Value) { return self.SubtractByInt64(Value); }
+        friend AltDec operator-(const AltDec& self, const Int64& Value) { return self.SubtractByInt64(Value); }
         friend AltDec operator-(const AltDec& self, const unsigned int& Value) { return self.SubtractByUInt(Value); }
-        friend AltDec operator-(const AltDec& self, const unsigned long long& Value) { return self.SubtractByUInt64(Value); }
+        friend AltDec operator-(const AltDec& self, const UInt64& Value) { return self.SubtractByUInt64(Value); }
 		
         friend AltDec operator-(const signed int& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
-        friend AltDec operator-(const signed long long& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
+        friend AltDec operator-(const Int64& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
         friend AltDec operator-(const unsigned int& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
-        friend AltDec operator-(const unsigned long long& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
+        friend AltDec operator-(const UInt64& lValue, const AltDec& rValue) { return ((AltDec)lValue).SubtractBy(rValue); }
 
         friend AltDec operator-(const AltDec& self, const signed char& Value) { return self.SubtractByInt8(Value); }
         friend AltDec operator-(const AltDec& self, const signed short& Value) { return self.SubtractByInt16(Value); }
@@ -1789,9 +1789,9 @@ public:
         /// <param name="Value">The right side value.</param>
         /// <returns>AltDec</returns>
         friend AltDec& operator-=(AltDec& self, const signed int& Value) { return self.IntSubOp(Value); }
-        friend AltDec& operator-=(AltDec& self, const signed long long& Value) { return self.Int64SubOp(Value); }
+        friend AltDec& operator-=(AltDec& self, const Int64& Value) { return self.Int64SubOp(Value); }
         friend AltDec& operator-=(AltDec& self, const unsigned int& Value) { return self.UIntSubOp(Value); }
-        friend AltDec& operator-=(AltDec& self, const unsigned long long& Value) { return self.UInt64SubOp(Value); }
+        friend AltDec& operator-=(AltDec& self, const UInt64& Value) { return self.UInt64SubOp(Value); }
 
         friend AltDec& operator-=(AltDec& self, const signed char& Value) { return self.Int8SubOp(Value); }
         friend AltDec& operator-=(AltDec& self, const signed short& Value) { return self.Int16SubOp(Value); }
